fix(api): value out of range

+++ LoliaFrpClient/Core/User/Info/InfoGetResponse_data.cs
@@ -71,9 +71,9 @@ namespace LoliaFrpClient.Core.User.Info
         /// <summary>The today_checked property</summary>
         public bool? TodayChecked { get; set; }
         /// <summary>The traffic_limit property</summary>
-        public int? TrafficLimit { get; set; }
+        public long? TrafficLimit { get; set; }
         /// <summary>The traffic_used property</summary>
-        public int? TrafficUsed { get; set; }
+        public long? TrafficUsed { get; set; }
         /// <summary>The username property</summary>
 #if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_1_OR_GREATER
 #nullable enable
@@ -120,8 +120,8 @@ namespace LoliaFrpClient.Core.User.Info
                 { "max_tunnel_count", n => { MaxTunnelCount = n.GetIntValue(); } },
                 { "role", n => { Role = n.GetStringValue(); } },
                 { "today_checked", n => { TodayChecked = n.GetBoolValue(); } },
-                { "traffic_limit", n => { TrafficLimit = n.GetIntValue(); } },
-                { "traffic_used", n => { TrafficUsed = n.GetIntValue(); } },
+                { "traffic_limit", n => { TrafficLimit = n.GetLongValue(); } },
+                { "traffic_used", n => { TrafficUsed = n.GetLongValue(); } },
                 { "username", n => { Username = n.GetStringValue(); } },
             };
         }
@@ -145,8 +145,8 @@ namespace LoliaFrpClient.Core.User.Info
             writer.WriteIntValue("max_tunnel_count", MaxTunnelCount);
             writer.WriteStringValue("role", Role);
             writer.WriteBoolValue("today_checked", TodayChecked);
-            writer.WriteIntValue("traffic_limit", TrafficLimit);
-            writer.WriteIntValue("traffic_used", TrafficUsed);
+            writer.WriteLongValue("traffic_limit", TrafficLimit);
+            writer.WriteLongValue("traffic_used", TrafficUsed);
             writer.WriteStringValue("username", Username);
             writer.WriteAdditionalData(AdditionalData);
         }
